+++ tests/server/useTimeoutEffect_test.ts
@@ -1,0 +1,19 @@
+import { renderHook } from '@testing-library/react-hooks/server'
+
+import { useTimeoutEffect } from '@/useTimeoutEffect'
+
+describe('useTimeoutEffect', () => {
+  it('should not throw error', () => {
+    const callback = jest.fn()
+    const { result } = renderHook(() =>
+      useTimeoutEffect(
+        {
+          callback
+        },
+        []
+      )
+    )
+
+    expect(result.error).toBeUndefined()
+  })
+})